+++ src/layer/softplus.cpp
@@ -37,7 +37,7 @@ int Softplus::forward_inplace(Mat& bottom_top_blob, const Option& opt) const
         float* ptr = bottom_top_blob.channel(q);
         for (int i = 0; i < size; i++)
         {
-            ptr[i] = log(exp(ptr[i]) + 1.0f);
+            ptr[i] = logf(expf(ptr[i]) + 1.0f);
         }
     }
