+++ src/layer/vulkan/multiheadattention_vulkan.cpp
@@ -47,7 +47,7 @@ int MultiHeadAttention_vulkan::create_pipeline(const Option& opt)
 {
     const int embed_dim_per_head = embed_dim / num_heads;
     {
-        const float inv_sqrt_embed_dim_per_head = 1.f / sqrt(embed_dim_per_head);
+        const float inv_sqrt_embed_dim_per_head = 1.f / sqrtf(embed_dim_per_head);
 
         q_gemm = ncnn::create_layer(ncnn::LayerType::Gemm);
         q_gemm->vkdev = vkdev;
