+++ src/layer/x86/deformableconv2d_pack4to1.h
@@ -100,8 +100,8 @@ static void deformableconv2d_pack4to1_sse(const std::vector<Mat>& bottom_blobs, 
                         int v4_pos = 0;
                         if (cond)
                         {
-                            int h_low = floor(h_im);
-                            int w_low = floor(w_im);
+                            int h_low = (int)floorf(h_im);
+                            int w_low = (int)floorf(w_im);
                             int h_high = h_low + 1;
                             int w_high = w_low + 1;
